// UI/greatToolsFrontend/src/app/models/festivalModal.ts
export interface Festival {
  orderId: string;
  festName: string;
  audio: string;
  name: string;
  mainImg: string;
  movingImgSide: string;
  leftParda: string;
  rightParda: string;
  touchToShowImg: string;
  backgroundColor: string;
  shayari: string[];
  movingMessage: string;
  extraImg1: string;
  extraImg2?: string;
  extraImg3?: string;
}